feat: Show a menu for tile selection
Before this, a right-ctrl + right-click was just applying the first preselected tile on that surface.
Now, a menu pops up with all preselected tiles for that type of surface. The user can choose his tile, and click OK.
This applies the tile on the surface, and saves this choice in DB (the ..._choices table).