fix: Create SQL tables based on Streaming event
The building and default materials SQL tables need to be created after the building is loaded. I used to count the number of children of 'Root', but this was not robust.
This new technique is now listening to the 'stateChanged' event from UIStateManager in order to detect the building load up.