New menu that shows the surfaces status
For each room, a set of tilable surfaces is identified. The user needs to choose (or validate) a tile for each  one of these surfaces. This menu helps showing which surfaces have been validated or not via a color code (green or red for the moment).
Once all the surfaces of a room have been validated, the room itself is considered validated (and its color changes as well).