Store and use surfaces/rooms validity in DB
When a user clicks a surface valid (or invalid), this is now getting stored in DB, inside the _validatedSurfaces table.
Every time a surface's validity status changes, a check is also performed in order to deduce wether or not the whole room can be considered valid or not.
Toggles and title colors of the PerRoomListMenu are adapted conequently.